server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests
pydantic-settings
stripe
stytch
dspy
black
pytz
boto3
werkzeug
sqlalchemy
psycopg2-binary